testcases/requirements.txt
absl-py==0.1.13
allure-pytest==2.5.2
allure-python-commons==2.5.2
appnope==0.1.0
args==0.1.0
asn1crypto==0.24.0
astor==0.6.2
atomicwrites==1.2.1
attrs==17.4.0
autobahn==18.3.1
Automat==0.6.0
backports.tempfile==1.0rc1
backports.weakref==1.0rc1
beautifulsoup4==4.6.0
bleach==1.5.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
clint==0.5.1
clyent==1.2.1
constantly==15.1.0
cryptography==2.2.2
cycler==0.10.0
decorator==4.0.11
dlib==19.9.0
entrypoints==0.2.2
Flask==0.12.1
gast==0.2.0
grpcio==1.10.0
-e git+https://github.com/openai/gym.git@5f8d1fc1c6ea8a7dab44fcdab8a4ac1c24ba6759#egg=gym
h5py==2.7.1
html5lib==0.9999999
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.6.0
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.5.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==0.23
marshmallow==2.13.5
matplotlib==2.2.2
mistune==0.7.4
more-itertools==4.3.0
nbconvert==5.1.1
nbformat==4.3.0
nbpresent==3.0.0
nltk==3.2.5
numpy==1.13.3
olefile==0.44
opencv-python==3.2.0.6
pandas==0.20.3
pandocfilters==1.4.1
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
pluggy==0.7.1
prompt-toolkit==1.0.14
protobuf==3.5.2.post1
ptyprocess==0.5.1
py==1.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pyglet==1.2.4
Pygments==2.2.0
pyOpenSSL==17.5.0
pyparsing==2.1.4
pytest==3.8.1
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
qtconsole==4.3.0
requests==2.18.4
requests-toolbelt==0.7.1
scikit-learn==0.19.0
scipy==0.19.1
selenium==3.5.0
service-identity==17.0.0
shortuuid==0.5.0
simplegeneric==0.8.1
six==1.11.0
tabulate==0.7.7
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
terminado==0.6
testpath==0.3
tflearn==0.3.2
Theano==1.0.1
tornado==4.4.2
tqdm==4.11.2
traitlets==4.3.2
txaio==2.9.0
urllib3==1.22
utils==0.9.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.1
widgetsnbextension==2.0.0
wordcloud==1.4.1
zope.interface==4.4.3
